Seperate xcpretty helper into new file
