+++ xcodecommand/xcpretty_helper.go
